Set up some initial classes
TicTacToe is the game itself and hold state of
current game. Rounds, current, score, player turn

BoardUI will keep track of what moves have been
made and what spaces are open as well as print the
current state of the gameboard

Player will capture players names and will state
individual scores, and moves

Messages is currently being implemented as a
module but might be moved inside TicTacToe. Could
be better for the game to display messages, not
the BoardUI or players classes.