Add initial game logic without game end trigger
TicTacToe Class
TicTacToe::game_data runs a series of prompts that
returns an array of Objects needed before a game
is started. Logic is that, we can not start a
game without players and we can not initialize
players without names and a 'best of' series
number. The class method allows for that. Hard
code player color and icons.

BoardUI Class
Added logic to print the board, route user choices
through error checking and into the gameboard
instance variable array.

Player Class
Basic player set up. Name and hard-coded icons
that are colorized from gem.

Need to work on checking for the winner and
ending the current round, then checking if that
score gave the round-winner the permanent lead in
the overal series number. Example. Player wins 4
rounds in best of 7 series, total game should end.

Removed Messages module, wasn't necessary to
organization. Those messages seemed to be the job
of the game itself, TicTacToe.